requirements.txt
PyPDF2
streamlit
PIL
pandas
requests
azure.cognitiveservices.speech
dotenv